requirements.txt
Django==3.2.3
djangorestframework==3.11.1
django-cors-headers==3.7.0
psycopg2
flake8
djangorestframework-jwt==1.11.0
django-allauth==0.44.0
django-rest-auth==0.9.5
Pillow
